reuse the same connection to avoid deadlock (?)

+++ router.py
@@ -21,37 +21,43 @@ class Router:
         self.replica_costs = np.full(self.num_replicas, float('inf'), dtype=np.float32)
         self.routes = [-1 for _ in queries]
 
-    def _evaluate_cost(self, configurations):
+    def _evaluate_cost(self, configurations: list | None):
         try:
             for i_rep, replica in enumerate(self.replicas):
                 #print(f'* benchmarking on replica {i_rep + 1} of {len(self.replicas)}')
                 # here, we actually want to open a new connection, so as to not interfere with
                 # the existing virtual indexes in our main replica connection
-                with psycopg.connect(replica.connection_string()) as conn:
-                    with conn.cursor() as cur:
-                        indexes_required = 0
+                conn = replica.connection()
+                with conn.cursor() as cur:
+                    indexes_required = 0
 
+                    if configurations is not None:
+                        cur.execute('SELECT hypopg_reset();')
                         for config in configurations[i_rep]:
                             table = config[0]
                             columns = config[1]
                             indexes_required += 1
                             #print(f'creating index {indexes_required} : {table}')
-                            creation_string = 'CREATE INDEX candidate_index_%d ON %s (%s);' % (indexes_required, table, ', '.join(columns))
+                            creation_string = 'CREATE INDEX candidate_index_%d ON %s (%s)' % (indexes_required, table, ', '.join(columns))
                             cur.execute('SELECT indexrelid FROM hypopg_create_index($$%s$$);' % creation_string)
-                        
-                        REGEX = 'cost=([0-9]+\\.[0-9]+)'
+                    
+                    REGEX = 'cost=([0-9]+\\.[0-9]+)'
 
-                        for idx, query in enumerate(self.queries):
-                            #print(f'estimating query {idx + 1} cost of {len(self.queries)}')
-                            cur.execute('EXPLAIN %s;' % query)
-                            if after_timing := re.search(REGEX, cur.fetchone()[0], re.IGNORECASE):
-                                self.times[i_rep][idx] = float(after_timing.group(1))
-                        
+                    for idx, query in enumerate(self.queries):
+                        #print(f'estimating query {idx + 1} cost of {len(self.queries)}')
+                        cur.execute('EXPLAIN %s' % query)
+                        if after_timing := re.search(REGEX, cur.fetchone()[0], re.IGNORECASE):
+                            self.times[i_rep][idx] = float(after_timing.group(1))
+                    
+                    if configurations is not None:
                         cur.execute('SELECT hypopg_reset();')
+                    
+                    conn.commit()
 
         except Exception as err:
             print('got an exception in the database connection')
             print(err)
+            conn.rollback()
     
     def _evaluate_exe(self, configurations):
         try:
@@ -61,12 +67,13 @@ class Router:
                 with conn.cursor() as cur:
                     indexes_required = 0
 
-                    for config in configurations[i_rep]:
-                        table = config[0]
-                        columns = config[1]
-                        indexes_required += 1
-                        #print(f'creating index {indexes_required} : {table}')
-                        cur.execute('CREATE INDEX candidate_index_%d ON %s (%s);' % (indexes_required, table, ', '.join(columns)))
+                    if configurations is not None:
+                        for config in configurations[i_rep]:
+                            table = config[0]
+                            columns = config[1]
+                            indexes_required += 1
+                            #print(f'creating index {indexes_required} : {table}')
+                            cur.execute('CREATE INDEX candidate_index_%d ON %s (%s);' % (indexes_required, table, ', '.join(columns)))
                     
                     for idx, query in enumerate(self.queries):
                         #print(f'testing query {idx + 1} of {len(self.queries)}')
@@ -76,15 +83,17 @@ class Router:
 
                         self.times[i_rep][idx] = toc - tic
                     
-                    while indexes_required > 0:
-                        cur.execute('DROP INDEX candidate_index_%d;' % indexes_required)
-                        indexes_required -= 1
+                    if configurations is not None:
+                        while indexes_required > 0:
+                            cur.execute('DROP INDEX candidate_index_%d;' % indexes_required)
+                            indexes_required -= 1
                     
                     conn.commit()
 
         except Exception as err:
             print('got an exception in the database connection')
             print(err)
+            conn.rollback()
     
     def parse_state_matrix(self, state):
         parsed_config = []
@@ -101,8 +110,18 @@ class Router:
         
         return parsed_config
     
-    def evaluate(self, configurations):
-        configurations = self.parse_state_matrix(configurations)
+    def evaluate(self, configurations: np.ndarray | None = None):
+        '''
+        Evaluates the cost of every query across the various replicas,
+        and updates the routeing table and the minimum cost of processing
+        each query.
+
+        :param configurations:  the current state array, or `None` if the
+                                information is already persisted in the
+                                connection object.
+        '''
+        if configurations is not None:
+            configurations = self.parse_state_matrix(configurations)
         for replica in self.replicas:
             replica.drop_all_indexes(self.tables, 'exe')
         self.profiler.time_in('database.route')
